+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Bumped `facebook-ios-sdk` to 18.1.0 [SDKS-5160]
 
 #### Fixed
+- Fixed `QRCodeCollector` not preserving the complete QR code data URI in `content` [SDKS-5299]
 - Fixed `OidcWebClient` `.authSession` and `.ephemeralAuthSession` not completing for Universal Link (https) redirect URIs [SDKS-5239]
 - Fixed FIDO registration/authentication not launching automatically when the DaVinci form's `trigger` property is not `BUTTON` [SDKS-4552]
 - Fixed 5xx AM responses with a parseable error body being misclassified as `FailureNode` instead of `ErrorNode`, diverging from Android [SDKS-5358]